Make tagging read messages more explicit.
The user may mark the current message as read and go to the next or
previous message with 'J' or 'K', or toggle the unread tag with the 'N'
key.  This is necessary so the user can browse around a thread without
affecting the unread tag unless they mean to.