Improve framework for handling screen resizes between views.
This uses the same technique as version types.  When a screen is
destroyed, we leave a forwarding pointer to the new screen.  When we
return to a caller which has a copy of the old screen, it can then
fast-forward to the latest version of the screen, and adapt to a new
screen size as necessary.