Improve handling of multipart/* parts in pager.
Now we show multipart/* head lines unless (1) we know it is not useful
to show, and (2) there are no alternatives for this current part.

In particular, I saw a message with structure:

  multipart/alternative
  + text/plain
  + multipart/related
    + text/html

Cycling alternatives on the text/plain part should show
a multipart/related part containing a text/html part.
The text/html part should have no alternatives.