Better history behaviour for open part / URL commands.
Use the variant of text entry which starts off editing the last string
in the history, and preseed the configured or default command into the
history.  Previously we started the text entry with an initial string
equal to the last string in the history, so pressing up to go to the
previous history entry would appear to have no effect.

Use separate histories for the two commands as they can now be
configured differently.