thread/pager: Initialise search string to last-entered string from any view.
If you search for "aaa" in the index view, open a thread, search for
"bbb" within that thread, return to the index view, open another thread
(or the same thread), then press 'n', previously we would try to search
for "aaa" as it is the active search string in the index view.
The new behaviour is to search for "bbb" as it is the last-entered
search string in any view.