text_entry: Make path completion work more smoothly.
If completion selects the only possible choice, which is a directory
name (it ends with "/"), then the next time completion is invoked,
generate the names of files under that directory.

Previously, the user would have to first another character
(or otherwise cause the completion choices to be cleared) before
pressing Tab.