thread/pager: Prevent exclusions due to tags searched for by the user.
Say the user wants to view messages in a thread that would normally be
excluded by search.tag_exclude, e.g. spam. The user can explicitly
search for that tag in the index view. When opening the thread, we pass
the set of tags that were matched for the thread to notmuch search
(currently even tags not in search.tag_exclude), preventing notmuch
search from honouring exclusions for tags that the user explicitly
searched for in the index view, but still honouring other exclusions.