Use application/octet-stream for text attachments with non UTF-8 content.
This works around an issue with resuming a postponed message containing
a text/* attachment with non UTF-8 content. notmuch show does not
convert the text content to UTF-8 as intended, nor does it provide the
correct charset. To avoid that encountering that situation, assign the
media type application/octet-stream to any text attachment that contains
non UTF-8 content.