Add thread_ordering to config options
Although bower provides both threaded and flat views of a thread, the
default that appears when I open a new thread had been hardcoded to
be the threaded view. Since that makes some of my threads run off the
right side of my screen, I found myself almost automatically pressing
'O' after opening a message to view. This commit adds a config option
to customize the default view while retaining the current behaviour
if the option is not set.