Redraw screen after manual multipart toggle
Since toggling multipart/alternative generation changes the
displayed message, it is necessary to redraw the screen to
make those changes transparent to the user. The easiest way
to go seemed to be reusing the resize action. Unfortunately
I couldn't find a way to retain the info message yet, but
proper preview seemed more important. I guess we don't really
need this message anyway as the multipart header shows up in
preview.