Fix postpone/resume
Postpone/resume was broken by the first commit on the alt_html_filter
feature branch. Basically because the postponed message included
the html alternative which the resume part couldn't handle. This
commit fixes the issue by always using text/plain only for postponed
messages. Since the html part is generated automatically, this seems
to be the preferable way of postponing messages anyway.