Disable 'o' binding for toggle-expanded
This makes moving 'z' and 'Z' to the base pager a lot easier as
it decouples functionality. It may be easy to add this again once
both toggle_content and open_part are part of the shared pager,
or it might just be enough to have the 'z' binding available.