Move 's' and 'o' to plain pager
This change required major refactoring. Among other things, I
moved the prompts for save_part and open_part out to a separate
module (prompt_external) which is imported both in pager and in
thread_pager. I also moved quite a bit of the pager implementation
into its interface and duplicated a function for getting notmuch
command line arguments. In summary, I may not have made the code
as clean as it should be, but I got a somewhat working solution.