Improve filename suggestion for empty MessageId
Since the MessageId is set to the empty string for newly composed
messages, simply adding ".part" to that doesn't seem like the best
option anymore. Among other things because ".part" would be a
hidden file, which doesn't seem like a good default. This commit
simply set the suggested filename to "message.part" if MessageId
is empty.