Change header env variable naming for alt_html_filter
As discussed on GitHub, it would be nice to have some consistency
with `PART_CHARSET` and `PART_CONTENT_TYPE`, which are exposed to
other filter commands. I went with the following variables in the
end:

    MAIL_FROM
    MAIL_TO
    MAIL_CC
    MAIL_BCC
    MAIL_SUBJECT
    MAIL_REPLY_TO
    MAIL_IN_REPLY_TO

`MAIL` has a reasonable size for a prefix, I think, and since we
are talking about emails that consist of multiple parts I also
felt the distinction between `MAIL` and `PART` made sense.