Increase max thread pane size after a certain point
This is a proposal to solve the thread pane display on hidpi
screens. Cf. <https://github.com/wangp/bower/issues/60>. It
will use the previous maximum of 8 rows for the thread pane
for terminal sizes of up to 40 rows. For larger terminals it
will use 1/5 of the terminal size as the maximum. As it happens,
8/40 = 1/5, which ensures a seamless switch between the two.
Switching back to relative size at a certain point removes all
hard limits on the maximum thread pane size.