Use 'N' for next search result reverse direction
I realise that this is a little different from how most programs
handle the 'n' and 'N' keys. Most programs use 'n' for *continue
search in the direction it was started* and 'N' for *search in the
other direction than the one specified while starting the search*.
The only other program I know of that uses the keybindings proposed
here is <https://github.com/martanne/vis> (which also happens to be
my text editor of choice, so I may be biassed). Simply using 'n' for
search forward and 'N' for search reverse and ignoring the initial
search direction makes the implementation more straightforward,
however, since it requires less state. I also find it easier to use
myself as I don't have to remember the initial search direction
either. As mentioned, this may be personal preference, of course.