Indicate attachments in hidden alternative part(s).
Apple Mail produces messages with the structure:

    multipart/alternative
     ├─ text/plain
     └─ multipart/mixed
         ├─ text/html
         └─ application/pdf

When the text/plain part is shown, the user will not see that there is
an attachment hidden within the alternative part. To alleviate this
problem, next to the hint "z for alternative", display the number of
attachments in the hidden part(s), if any.

Fixes #111.