Allow manual overrides of the date header field
Previously, the date field was always updated when sending the
message. While this behaviour is quite reasonable in most cases,
there might also be moments when specifying a date different from
the system's current local time might come in handy. This commit
introduces the ability to override the date field by simply including
it in the temporary message. If a non-empty date header is included
in the temporary message, bower will use the user-supplied value
when sending the email. If no date header is included (the default)
or if the date header is empty (i. e. consists only of zero or more
spaces) bower will continue to use the system's current local time
to automatically provide a value for the date field.

This change follows the unix philosophy of always giving users
enough rope to hang themselves; and then to add a couple of feet
more just in case. I. e. a user-supplied date value is used as-is,
without any kinds of integrity checks. This gives users quite some
freedom when using manual overrides, but it also presumes that they
know what they are doing. The date override is also not documented
(yet), but I believe that the behaviour is quite intuitive for the
kinds of users who would want to use it.

Note that bower also automatically sets a date when postponing
messages. This date will currently also appear in the header when
the postponed message is resumed. In order to only include this
date if it was explicitly specified by the user, we would need to
introduce some means of preserving bower-specific metadata across
the postpone-resume cycle. This change will be left for a later
commit.