Retain date override state across postpone-resume
This introduces a custom header field called X-Bower-Metadata.
X-Bower-Metadata will be written into the header of postponed
messages if needed and it will be read back and removed from the
header when parsing postponed messages. The custom header field is
set up in a way that it could potentially store any number of boolean
options and maybe even arbitrary key-value-pairs. Its structure is a
list of strings separated by semicolons. Currently, all values except
for 'retain_date' are ignored. The value 'retain_date' will be stored
in X-Bower-Metadata if the date header was manually set by the user.

If X-Bower-Metadata does not specify 'retain_date', the date header
will be cleared upon opening a postponed message. This commit thus
restores the previous behaviour for resuming postponed messages when
the manual date override is not used, resolving the issue introduced
in commit c4521186.